bootstrap.sh: Fix SC2115
If somehow TMP_DIR is not set, it will error out like this:

    bash: flarp: parameter null or not set

Shellcheck Warning:

    In bootstrap-scripts/bootstrap.sh line 242:
        [ -d "$TMP_DIR" ] && rm -rf "$TMP_DIR/"
                                    ^---------^ SC2115 (warning): Use "${var:?}" to ensure this never expands to / .

    For more information:
      https://www.shellcheck.net/wiki/SC2115 -- Use "${var:?}" to ensure this nev...

+++ bootstrap-scripts/bootstrap.sh
@@ -239,7 +239,7 @@ if [ ! -d "/Applications/Xcode.app" ]; then
       sudo mv ./Xcode.app /Applications/
       popd
     fi
-    [ -d "$TMP_DIR" ] && rm -rf "$TMP_DIR/"
+    [ -d "$TMP_DIR" ] && rm -rf "${TMP_DIR:?}/"
   else
     hdiutil attach "$XCODE_DMG"
     export __CFPREFERENCES_AVOID_DAEMON=1
